automatically fix ivcurves

+++ diode_cap_buck_behavior/data_solar/convert_ivcurve.py
@@ -30,7 +30,9 @@ paths = [
     path_here / "LED_035pc.h5",
 ]
 
-for path in paths:
+cutoff_bin = list(range(90,50,-2))
+
+for iter, path in enumerate(paths):
     with Reader(path, verbose=False) as reader:
         samples = reader.get_window_samples()
         repetitions = int(reader.ds_time.shape[0] / samples)
@@ -46,9 +48,15 @@ for path in paths:
         curve_v = cal.voltage.raw_to_si(curve_v) / repetitions
         curve_i = cal.current.raw_to_si(curve_i) / repetitions
 
+    # fix spike at t0 and begin of current-curve
+    for _j in range(cutoff_bin[iter]-1,-1,-1):
+        curve_i[_j] = max(0.0, float((1+1/4)*curve_i[_j+1] - curve_i[_j+5]/4))
+
     separator = "; "
     csv_name = path.stem.split(".")[0]
     csv_path = path.with_stem(csv_name).with_suffix(".ivcurve.csv")
+    if csv_path.exists():
+        continue
     with csv_path.open("w", encoding="utf-8-sig") as csv_file:
         csv_file.write(separator.join(["Time [s]", "Voltage [V]", "Current [A]"]) + "\n")
         for idx in range(samples):
